relocate genScript to Wallet.Gen and set cardano-addresses-3.2.0 in stack
