+++ lib/core/test/unit/Cardano/Wallet/Primitive/ScriptsSpec.hs
@@ -10,7 +10,6 @@
 
 module Cardano.Wallet.Primitive.ScriptsSpec
     ( spec
-    , genScript
     ) where
 
 import Prelude
@@ -19,6 +18,8 @@ import Cardano.Address.Derivation
     ( XPub )
 import Cardano.Address.Script
     ( KeyHash (..), Script (..), ScriptHash, toScriptHash )
+import Cardano.Wallet.Gen
+    ( genScript )
 import Cardano.Wallet.Primitive.AddressDerivation
     ( Depth (..)
     , DerivationType (..)
@@ -63,23 +64,13 @@ import Data.Set
     ( Set )
 import Data.Word
     ( Word32 )
-import Numeric.Natural
-    ( Natural )
 import Test.Hspec
     ( Spec, describe, it )
 import Test.QuickCheck
     ( Arbitrary (..)
-    , Gen
-    , Positive (..)
     , Property
-    , arbitrarySizedNatural
     , choose
-    , elements
-    , oneof
     , property
-    , scale
-    , shrinkIntegral
-    , sized
     , vectorOf
     , (.&&.)
     , (===)
@@ -311,51 +302,6 @@ scriptKeyHashesInMap script' accXPub' s =
 {-------------------------------------------------------------------------------
                                 Arbitrary Instances
 -------------------------------------------------------------------------------}
-
-instance Arbitrary Natural where
-    shrink = shrinkIntegral
-    arbitrary = arbitrarySizedNatural
-
-genScript :: [KeyHash] -> Gen (Script KeyHash)
-genScript keyHashes = scale (`div` 3) $ sized scriptTree
-    where
-        scriptTree 0 = oneof
-            [ RequireSignatureOf <$> elements keyHashes
-            , ActiveFromSlot <$> arbitrary
-            , ActiveUntilSlot <$> arbitrary
-            ]
-        scriptTree n = do
-            Positive m <- arbitrary
-            let n' = n `div` (m + 1)
-            scripts' <- vectorOf m (scriptTree n')
-            let hasTimelocks = \case
-                    ActiveFromSlot _ -> True
-                    ActiveUntilSlot _ -> True
-                    _ -> False
-            let scriptsWithValidTimelocks = case L.partition hasTimelocks scripts' of
-                    ([], rest) -> rest
-                    ([ActiveFromSlot s1, ActiveUntilSlot s2], rest) ->
-                        if s2 <= s1 then
-                            rest ++ [ActiveFromSlot s2, ActiveUntilSlot s1]
-                        else
-                            scripts'
-                    ([ActiveUntilSlot s2, ActiveFromSlot s1], rest) ->
-                        if s2 <= s1 then
-                            rest ++ [ActiveFromSlot s2, ActiveUntilSlot s1]
-                        else
-                            scripts'
-                    ([ActiveFromSlot _], _) -> scripts'
-                    ([ActiveUntilSlot _], _) -> scripts'
-                    (_,rest) -> rest
-            case fromIntegral (L.length (filter (not . hasTimelocks) scriptsWithValidTimelocks)) of
-                0 -> scriptTree 0
-                num -> do
-                    atLeast <- choose (1, num)
-                    elements
-                        [ RequireAllOf scriptsWithValidTimelocks
-                        , RequireAnyOf scriptsWithValidTimelocks
-                        , RequireSomeOf atLeast scriptsWithValidTimelocks
-                        ]
 
 prepareVerKeys
     :: ShelleyKey 'AccountK XPub
