+++ stack.yaml
@@ -15,8 +15,6 @@ packages:
 - lib/test-utils
 - lib/shelley
 - lib/strict-non-empty-containers
-- ../cardano-addresses/command-line
-- ../cardano-addresses/core
 
 extra-deps:
 # Miscellaneous
@@ -67,12 +65,12 @@ extra-deps:
     - persistent-sqlite
     - persistent-template
 
-# cardano-addresses-3.1.0
-#- git: https://github.com/input-output-hk/cardano-addresses
-#  commit: 9c38189c3e0e011a305ddcd1219e50a0731e99ea
-#  subdirs:
-#    - command-line
-#    - core
+# cardano-addresses-3.2.0
+- git: https://github.com/input-output-hk/cardano-addresses
+  commit: 115174cc451d3fc6b90ce61782f841e51f271c3d
+  subdirs:
+    - command-line
+    - core
 
 # cardano-transactions master branch
 - git: https://github.com/input-output-hk/cardano-transactions
